requirements.txt
requests>=2.28.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
jupyter>=1.0.0
numpy>=1.23.0
python-dotenv>=1.0.0

